requirements.txt
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.5
opencv-contrib-python==4.11.0.86
pillow==11.2.1
setuptools==70.2.0
sympy==1.13.3
torch==2.7.0+cpu
torchaudio==2.7.0+cpu
torchvision==0.22.0+cpu
typing_extensions==4.12.2
